Src folder
Moved project files into src/
