--- commands/General/modabuse.js
@@ -1,12 +1,0 @@
-import { MessageEmbed } from "discord.js"
-import { interactionReply } from "../../utils.js"
-import { config } from "../../index.js"
-
-export const description = "Funny meme about mod abuse"
-export function run (interaction) {
-	interactionReply(interaction, {
-		messageEmbed: new MessageEmbed()
-			.setColor(config.embedDesign.color.gold)
-			.setImage(config.generalLinks.modAbuseImage)
-	})
-}
